[FB対応]hooksファイルとcomponentsファイルにそれぞれを収納
